backend/requirements.txt
flask==3.0.2
flask-cors==4.0.0
python-dotenv==1.0.1
groq==0.4.2  # Latest Groq client
httpx==0.26.0  # Required by Groq
pymongo==4.6.1
python-dateutil==2.8.2
typing-extensions==4.9.0
requests==2.32.3
gunicorn==21.2.0  # For production deployment

# Python version
# python>=3.12.0
